Fixed bugs related to variable rename feature. Updated delete button to bring up pop-up warning. Changed the varLoader variable rename feature so name is updated as user types. Updated save dialog so that selection/deselection is off when the dialog is open. Fixed minor bug where deleting a variable would not deselect it.
